+++ google2/src/main/scala/org/broadinstitute/dsde/workbench/google2/GoogleKmsInterpreter.scala
@@ -1,6 +1,6 @@
 package org.broadinstitute.dsde.workbench.google2
 
-import cats.effect.{ContextShift, Resource, Sync}
+import cats.effect.{Blocker, ContextShift, Resource, Sync}
 import cats.implicits._
 import com.google.api.gax.core.FixedCredentialsProvider
 import com.google.auth.oauth2.ServiceAccountCredentials
@@ -8,7 +8,6 @@ import com.google.cloud.kms.v1.CryptoKey.CryptoKeyPurpose
 import com.google.cloud.kms.v1.{CryptoKey, CryptoKeyName, KeyManagementServiceClient, KeyManagementServiceSettings, KeyRing, KeyRingName, LocationName}
 import com.google.iam.v1.{Binding, Policy}
 import com.google.protobuf.{Duration, Timestamp}
-import io.chrisdavenport.linebacker.Linebacker
 import org.broadinstitute.dsde.workbench.model.google.GoogleProject
 
 import scala.collection.JavaConverters._
@@ -20,7 +19,7 @@ import scala.language.higherKinds
   * created by mtalbott on 1/18/19
   */
 
-private[google2] class GoogleKmsInterpreter[F[_]: Sync: ContextShift: Linebacker](client: KeyManagementServiceClient) extends GoogleKmsService[F] {
+private[google2] class GoogleKmsInterpreter[F[_]: Sync: ContextShift](client: KeyManagementServiceClient, blocker: Blocker) extends GoogleKmsService[F] {
   override def createKeyRing(project: GoogleProject, location: Location, keyRingId: KeyRingId): F[KeyRing] = {
     val locationName = LocationName.format(project.value, location.value)
     blockingF(Sync[F].delay[KeyRing] {
@@ -109,11 +108,11 @@ private[google2] class GoogleKmsInterpreter[F[_]: Sync: ContextShift: Linebacker
     })
   }
 
-  private def blockingF[A](fa: F[A]): F[A] = Linebacker[F].blockCS(fa)
+  private def blockingF[A](fa: F[A]): F[A] = blocker.blockOn(fa)
 }
 
 object GoogleKmsInterpreter {
-  def apply[F[_]: Sync: ContextShift: Linebacker](client: KeyManagementServiceClient): GoogleKmsInterpreter[F] = new GoogleKmsInterpreter[F](client)
+  def apply[F[_]: Sync: ContextShift](client: KeyManagementServiceClient, blocker: Blocker): GoogleKmsInterpreter[F] = new GoogleKmsInterpreter[F](client, blocker)
 
   def client[F[_]: Sync](pathToJson: String): Resource[F, KeyManagementServiceClient] =
     for {